Cleaned up error raising. Moved superion code into lib/hyperion
