+++ lib/hyperion/result_handling/dispatching_hyperion_result.rb
@@ -2,6 +2,8 @@ require 'hyperion/types/hyperion_result'
 require 'hyperion/result_handling/dispatch_dsl'
 
 class Hyperion
+  # PW: distinguishing between this and HyperionResult is of
+  # dubious value. Consider merging the two.
   class DispatchingHyperionResult < HyperionResult
     include DispatchDsl
   end
