+++ lib/hyperion/types/hyperion_error.rb
@@ -1,0 +1,6 @@
+class HyperionError < StandardError
+end
+
+def hyperion_raise(msg)
+  raise HyperionError, msg
+end
